+++ SentryTestUtils/SentryTestUtils-ObjC-BridgingHeader.h
@@ -8,6 +8,12 @@
 #    import "SentryUIViewControllerPerformanceTracker.h"
 #endif // SENTRY_HAS_UIKIT
 
+#import "SentryProfilingConditionals.h"
+
+#if SENTRY_TARGET_PROFILING_SUPPORTED
+#    import "SentryProfiler+Test.h"
+#endif // SENTRY_TARGET_PROFILING_SUPPORTED
+
 #import "PrivateSentrySDKOnly.h"
 #import "SentryAppState.h"
 #import "SentryClient+Private.h"
@@ -26,7 +32,6 @@
 #import "SentryNSTimerFactory.h"
 #import "SentryNetworkTracker.h"
 #import "SentryPerformanceTracker+Testing.h"
-#import "SentryProfiler+Test.h"
 #import "SentryRandom.h"
 #import "SentrySDK+Private.h"
 #import "SentrySDK+Tests.h"