+++ Tests/SentryProfilerTests/SentryProfilerSwiftTests.swift
@@ -30,6 +30,7 @@ class SentryProfilerSwiftTests: XCTestCase {
         lazy var dispatchFactory = TestDispatchFactory()
         var metricTimerFactory: TestDispatchSourceWrapper?
         lazy var timeoutTimerFactory = TestSentryNSTimerFactory()
+        let dispatchQueueWrapper = TestSentryDispatchQueueWrapper()
 
         let currentDateProvider = TestCurrentDateProvider()
 
@@ -63,11 +64,25 @@ class SentryProfilerSwiftTests: XCTestCase {
         }
 
         /// Advance the mock date provider, start a new transaction and return its handle.
-        func newTransaction(testingAppLaunchSpans: Bool = false) throws -> SentryTracer {
-            if testingAppLaunchSpans {
-                return try XCTUnwrap(hub.startTransaction(name: transactionName, operation: SentrySpanOperationUILoad) as? SentryTracer)
+        func newTransaction(testingAppLaunchSpans: Bool = false, automaticTransaction: Bool = false, idleTimeout: TimeInterval? = nil) throws -> SentryTracer {
+            let operation = testingAppLaunchSpans ? SentrySpanOperationUILoad : transactionOperation
+
+            if automaticTransaction {
+                return hub.startTransaction(
+                    with: TransactionContext(name: transactionName, operation: operation),
+                    bindToScope: false,
+                    customSamplingContext: [:],
+                    configuration: SentryTracerConfiguration(block: {
+                        if let idleTimeout = idleTimeout {
+                            $0.idleTimeout = idleTimeout
+                        }
+                        $0.dispatchQueueWrapper = self.dispatchQueueWrapper
+                        $0.waitForChildren = true
+                        $0.timerFactory = self.timeoutTimerFactory
+                    }))
             }
-            return try XCTUnwrap(hub.startTransaction(name: transactionName, operation: transactionOperation) as? SentryTracer)
+
+            return try XCTUnwrap(hub.startTransaction(name: transactionName, operation: operation) as? SentryTracer)
         }
 
         // mocking
@@ -243,10 +258,12 @@ class SentryProfilerSwiftTests: XCTestCase {
 
         func createConcurrentSpansWithMetrics() throws {
             XCTAssertFalse(SentryProfiler.isCurrentlyProfiling())
+            XCTAssertEqual(SentryProfiler.currentProfiledTracers(), UInt(0))
 
-            for _ in 0 ..< numberOfTransactions {
+            for i in 0 ..< numberOfTransactions {
                 let span = try fixture.newTransaction()
                 XCTAssertTrue(SentryProfiler.isCurrentlyProfiling())
+                XCTAssertEqual(SentryProfiler.currentProfiledTracers(), UInt(i + 1))
                 spans.append(span)
                 fixture.currentDateProvider.advanceBy(nanoseconds: 100)
             }
@@ -257,12 +274,14 @@ class SentryProfilerSwiftTests: XCTestCase {
                 try fixture.gatherMockedMetrics(span: span)
                 XCTAssertTrue(SentryProfiler.isCurrentlyProfiling())
                 span.finish()
+                XCTAssertEqual(SentryProfiler.currentProfiledTracers(), UInt(numberOfTransactions - i - 1))
 
                 try self.assertValidProfileData()
                 try self.assertMetricsPayload(metricsBatches: i + 1)
             }
             
             XCTAssertFalse(SentryProfiler.isCurrentlyProfiling())
+            XCTAssertEqual(SentryProfiler.currentProfiledTracers(), UInt(0))
         }
 
         try createConcurrentSpansWithMetrics()
@@ -401,6 +420,108 @@ class SentryProfilerSwiftTests: XCTestCase {
             options.profilesSampler = { _ in return -0.01 }
         }
     }
+
+    /// based on ``SentryTracerTests.testFinish_WithoutHub_DoesntCaptureTransaction``
+    func testProfilerCleanedUpAfterTransactionDiscarded_NoHub() throws {
+        XCTAssertEqual(SentryProfiler.currentProfiledTracers(), UInt(0))
+        func performTransaction() {
+            let sut = SentryTracer(transactionContext: TransactionContext(name: fixture.transactionName, operation: fixture.transactionOperation), hub: nil)
+            XCTAssertEqual(SentryProfiler.currentProfiledTracers(), UInt(0))
+            sut.finish()
+        }
+        performTransaction()
+        XCTAssertEqual(SentryProfiler.currentProfiledTracers(), UInt(0))
+        XCTAssertEqual(self.fixture.client?.captureEventWithScopeInvocations.count, 0)
+    }
+
+    /// based on ``SentryTracerTests.testFinish_WaitForAllChildren_ExceedsMaxDuration_NoTransactionCaptured``
+    func testProfilerCleanedUpAfterTransactionDiscarded_ExceedsMaxDuration() throws {
+        XCTAssertEqual(SentryProfiler.currentProfiledTracers(), UInt(0))
+        func performTransaction() throws {
+            let sut = try fixture.newTransaction(automaticTransaction: true)
+            XCTAssertEqual(SentryProfiler.currentProfiledTracers(), UInt(1))
+            fixture.currentDateProvider.advance(by: 500)
+            sut.finish()
+        }
+        try performTransaction()
+        XCTAssertEqual(SentryProfiler.currentProfiledTracers(), UInt(0))
+        XCTAssertEqual(self.fixture.client?.captureEventWithScopeInvocations.count, 0)
+    }
+
+    func testProfilerCleanedUpAfterInFlightTransactionDeallocated() throws {
+        XCTAssertEqual(SentryProfiler.currentProfiledTracers(), UInt(0))
+        func performTransaction() throws {
+            let sut = try fixture.newTransaction(automaticTransaction: true)
+            XCTAssertEqual(SentryProfiler.currentProfiledTracers(), UInt(1))
+            XCTAssertFalse(sut.isFinished)
+        }
+        try performTransaction()
+        XCTAssertEqual(SentryProfiler.currentProfiledTracers(), UInt(0))
+        XCTAssertEqual(self.fixture.client?.captureEventWithScopeInvocations.count, 0)
+    }
+
+    /// based on ``SentryTracerTests.testFinish_IdleTimeout_ExceedsMaxDuration_NoTransactionCaptured``
+    func testProfilerCleanedUpAfterTransactionDiscarded_IdleTimeout_ExceedsMaxDuration() throws {
+        XCTAssertEqual(SentryProfiler.currentProfiledTracers(), UInt(0))
+        func performTransaction() throws {
+            let sut = try fixture.newTransaction(automaticTransaction: true, idleTimeout: 1)
+            XCTAssertEqual(SentryProfiler.currentProfiledTracers(), UInt(1))
+            fixture.currentDateProvider.advance(by: 500)
+            sut.finish()
+        }
+        try performTransaction()
+        XCTAssertEqual(SentryProfiler.currentProfiledTracers(), UInt(0))
+        XCTAssertEqual(self.fixture.client?.captureEventWithScopeInvocations.count, 0)
+    }
+
+    /// based on ``SentryTracerTests.testIdleTimeout_NoChildren_TransactionNotCaptured``
+    func testProfilerCleanedUpAfterTransactionDiscarded_IdleTimeout_NoChildren() throws {
+        XCTAssertEqual(SentryProfiler.currentProfiledTracers(), UInt(0))
+        func performTransaction() throws {
+            let span = try fixture.newTransaction(automaticTransaction: true, idleTimeout: 1)
+            XCTAssertEqual(SentryProfiler.currentProfiledTracers(), UInt(1))
+            fixture.currentDateProvider.advance(by: 500)
+            fixture.dispatchQueueWrapper.invokeLastDispatchAfter()
+            XCTAssert(span.isFinished)
+        }
+        try performTransaction()
+        XCTAssertEqual(SentryProfiler.currentProfiledTracers(), UInt(0))
+        XCTAssertEqual(self.fixture.client?.captureEventWithScopeInvocations.count, 0)
+    }
+
+    /// based on ``SentryTracerTests.testIdleTransaction_CreatingDispatchBlockFails_NoTransactionCaptured``
+    func testProfilerCleanedUpAfterTransactionDiscarded_IdleTransaction_CreatingDispatchBlockFails() throws {
+        fixture.dispatchQueueWrapper.createDispatchBlockReturnsNULL = true
+        XCTAssertEqual(SentryProfiler.currentProfiledTracers(), UInt(0))
+        func performTransaction() throws {
+            let span = try fixture.newTransaction(automaticTransaction: true, idleTimeout: 1)
+            XCTAssertEqual(SentryProfiler.currentProfiledTracers(), UInt(1))
+            fixture.currentDateProvider.advance(by: 500)
+            span.finish()
+        }
+        try performTransaction()
+        XCTAssertEqual(SentryProfiler.currentProfiledTracers(), UInt(0))
+        XCTAssertEqual(self.fixture.client?.captureEventWithScopeInvocations.count, 0)
+    }
+
+#if os(iOS) || os(tvOS) || targetEnvironment(macCatalyst)
+    /// based on ``SentryTracerTests.testFinish_WaitForAllChildren_StartTimeModified_NoTransactionCaptured``
+    func testProfilerCleanedUpAfterTransactionDiscarded_WaitForAllChildren_StartTimeModified() throws {
+        XCTAssertEqual(SentryProfiler.currentProfiledTracers(), UInt(0))
+        let appStartMeasurement = fixture.getAppStartMeasurement(type: .cold)
+        SentrySDK.setAppStartMeasurement(appStartMeasurement)
+        fixture.currentDateProvider.advance(by: 1)
+        func performTransaction() throws {
+            let sut = try fixture.newTransaction(testingAppLaunchSpans: true, automaticTransaction: true)
+            XCTAssertEqual(SentryProfiler.currentProfiledTracers(), UInt(1))
+            fixture.currentDateProvider.advance(by: 499)
+            sut.finish()
+        }
+        try performTransaction()
+        XCTAssertEqual(SentryProfiler.currentProfiledTracers(), UInt(0))
+        XCTAssertEqual(self.fixture.client?.captureEventWithScopeInvocations.count, 0)
+    }
+#endif // os(iOS) || os(tvOS) || targetEnvironment(macCatalyst)
 }
 
 private extension SentryProfilerSwiftTests {